Consolidate overlapping audio engine tests
Merged `test_audio_engine_logic.py`, `test_audio_engine_dithering.py`, and `test_audio_error_handling.py` into `test_audio_engine.py` to organize and clean up the test suite. All tests pass locally.
